feat: add jwt

+++ src/services/posts.js
@@ -1,7 +1,7 @@
 import { Post } from '../db/models/post.js'
 
-export async function createPost(postModel) {
-  const post = new Post(postModel)
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
   return await post.save()
 }
 
@@ -31,14 +31,18 @@ export async function listPostsByTag(tags, options) {
   return await listPosts({ tags }, options)
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(
+  userId,
+  postId,
+  { title, author, contents, tags },
+) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
+    { _id: postId, author: userId },
     { $set: { title, author, contents, tags } },
     { new: true },
   )
 }
 
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
